♻️ refactor(contract): relocate REST @Resource surface to :server
