+++ server/src/jvmMain/kotlin/com/calypsan/listenup/server/routes/CollectionRoutes.kt
@@ -5,7 +5,7 @@ import com.calypsan.listenup.api.dto.CreateCollectionBody
 import com.calypsan.listenup.api.dto.SharePermission
 import com.calypsan.listenup.api.dto.ShareCollectionBody
 import com.calypsan.listenup.api.error.AppError
-import com.calypsan.listenup.api.resources.CollectionResources
+import com.calypsan.listenup.server.routes.resources.CollectionResources
 import com.calypsan.listenup.api.result.AppResult
 import com.calypsan.listenup.core.BookId
 import com.calypsan.listenup.core.CollectionId
