+++ sharedLogic/src/commonTest/kotlin/com/calypsan/listenup/konsist/ContractHasNoKtorClientRule.kt
@@ -5,9 +5,8 @@ import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldBeEmpty
 
 /**
- * `:contract` is the cross-platform wire/domain contract. It may use `io.ktor.resources`
- * (the `@Resource` REST-route surface) and `io.ktor.http`/`io.ktor.utils.io` (FileSource),
- * but it must NOT depend on the Ktor CLIENT transport — that coupling moved to the client
+ * `:contract` is the cross-platform wire/domain contract. It may use `io.ktor.http`/`io.ktor.utils.io`
+ * (FileSource), but it must NOT depend on the Ktor CLIENT transport — that coupling moved to the client
  * with ErrorMapper. Pin the boundary so a client-transport import can't creep back into :contract.
  */
 class ContractHasNoKtorClientRule :
